+++ src/OpenInApp.Common/Helpers/Dtos/ApplicationToOpenDto.cs
@@ -9,7 +9,7 @@ namespace OpenInApp.Common.Helpers.Dtos
         public string ExecutableFileToBrowseFor { get; set; }
         public InitialFolderType InitialFolderType { get; set; }//gregtt do we always look in \ProgFiles anyway, effectively making this redundant ?
         public string SecondaryFilePathSegment { get; set; }
-        public bool SecondaryFilePathSegmentHasMultipleYearNumberVersions { get; set; }
+        public bool SecondaryFilePathSegmentHasMultipleVersions { get; set; }
         public bool SeparateProcessPerFileToBeOpened { get; set; } //always true, ergo can delete this property ?
         public bool UseShellExecute { get; set; }
     }