+++ CHANGELOG.md
@@ -4,8 +4,8 @@ These are the changes to each version that has been released in the Visual Studi
 
 ## 1.1.x
 **2017-07-xx**
-- [x] Addition of Open In Atom ?????????????????????????????
-- [x] Addition of Open In SQL Server Management Studio  NEED TO SORT OUT PATH TO SSMS !!!!!!!!!!!!!!!!!!!!!!!!!!
+- [x] Addition of Open In Atom ?????
+- [x] Addition of Open In SQL Server Management Studio 2005/8/10/12/14/16/17,etc NEED TO SORT OUT PATH TO SSMS !!!!!!
 - [x] Addition of project nodes to TreeSize Free/Pro & WinDirStat
 - [x] Better VSMM descriptions for all apps